+++ cognite/client/data_classes/files.py
@@ -15,6 +15,8 @@ class FileMetadata(CogniteResource):
         mime_type (str): File type. E.g. text/plain, application/pdf, ..
         metadata (Dict[str, Any]): Custom, application specific metadata. String key -> String value. Limits: Maximum length of key is 32 bytes, value 512 bytes, up to 16 key-value pairs.
         asset_ids (List[int]): No description.
+        source_created_time (int): The timestamp for when the file was originally created in the source system.
+        source_modified_time (int): The timestamp for when the file was last modified in the source system.
         id (int): A JavaScript-friendly internal ID for the object.
         uploaded (bool): Whether or not the actual file is uploaded.  This field is returned only by the API, it has no effect in a post body.
         uploaded_time (int): The number of milliseconds since 00:00:00 Thursday, 1 January 1970, Coordinated Universal Time (UTC), minus leap seconds.
@@ -31,6 +33,8 @@ class FileMetadata(CogniteResource):
         mime_type: str = None,
         metadata: Dict[str, Any] = None,
         asset_ids: List[int] = None,
+        source_created_time: int = None,
+        source_modified_time: int = None,
         id: int = None,
         uploaded: bool = None,
         uploaded_time: int = None,
@@ -44,6 +48,8 @@ class FileMetadata(CogniteResource):
         self.mime_type = mime_type
         self.metadata = metadata
         self.asset_ids = asset_ids
+        self.source_created_time = source_created_time
+        self.source_modified_time = source_modified_time
         self.id = id
         self.uploaded = uploaded
         self.uploaded_time = uploaded_time
@@ -67,6 +73,8 @@ class FileMetadataFilter(CogniteFilter):
         created_time (Dict[str, Any]): Range between two timestamps.
         last_updated_time (Dict[str, Any]): Range between two timestamps.
         uploaded_time (Dict[str, Any]): Range between two timestamps.
+        source_created_time (Dict[str, Any]): Filter for files where the sourceCreatedTime field has been set and is within the specified range.
+        source_modified_time (Dict[str, Any]): Filter for files where the sourceModifiedTime field has been set and is within the specified range.
         external_id_prefix (str): The external ID provided by the client. Must be unique within the project.
         uploaded (bool): Whether or not the actual file is uploaded. This field is returned only by the API, it has no effect in a post body.
         cognite_client (CogniteClient): The client to associate with this object.
@@ -82,6 +90,8 @@ class FileMetadataFilter(CogniteFilter):
         created_time: Dict[str, Any] = None,
         last_updated_time: Dict[str, Any] = None,
         uploaded_time: Dict[str, Any] = None,
+        source_created_time: Dict[str, Any] = None,
+        source_modified_time: Dict[str, Any] = None,
         external_id_prefix: str = None,
         uploaded: bool = None,
         cognite_client=None,
@@ -94,6 +104,8 @@ class FileMetadataFilter(CogniteFilter):
         self.created_time = created_time
         self.last_updated_time = last_updated_time
         self.uploaded_time = uploaded_time
+        self.source_created_time = source_created_time
+        self.source_modified_time = source_modified_time
         self.external_id_prefix = external_id_prefix
         self.uploaded = uploaded
         self._cognite_client = cognite_client
@@ -123,6 +135,14 @@ class FileMetadataUpdate(CogniteUpdate):
     @property
     def asset_ids(self):
         return _ListFileMetadataUpdate(self, "assetIds")
+
+    @property
+    def source_created_time(self):
+        return _PrimitiveFileMetadataUpdate(self, "sourceCreatedTime")
+
+    @property
+    def source_modified_time(self):
+        return _PrimitiveFileMetadataUpdate(self, "sourceModifiedTime")
 
 
 class _PrimitiveFileMetadataUpdate(CognitePrimitiveUpdate):