+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Changes are grouped as follows
 - DatapointsAPI.insert now accepts a Datapoints object for easier copying
 - helper method `cognite.client.testing.mock_cognite_client()` for mocking CogniteClient
 - parent_id and parent_external_id to AssetUpdate class.
+- source_created_time and source_modified_time support for files
 
 ### Changed
 - assets.create() no longer validates asset hierarchy and sorts assets before posting. This functionality has been moved to assets.create_hierarchy().