+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Changes are grouped as follows
 - include_aggregate_names option on datapoints.retrieve_dataframe and Datapoints(List).to_pandas
 - datapoints.retrieve_dataframe_dict function, which returns {aggregate:dataframe} without adding aggregate names to columns
 - parent_id and parent_external_id to AssetUpdate class.
+- source_created_time and source_modified_time support for files
 
 ### Changed
 - assets.create() no longer validates asset hierarchy and sorts assets before posting. This functionality has been moved to assets.create_hierarchy().