+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Changes are grouped as follows
 - Separate read/write fields on data classes
 
 ## [Unreleased]
+### Added
 - source_created_time and source_modified_time support for files
 
 ## [1.1.0] - 2019-08-21